% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{calc_fgsea_w}
\alias{calc_fgsea_w}
\title{Calc GSEA}
\usage{
calc_fgsea_w(NMF_model, cutoff = 1e-04)
}
\arguments{
\item{NMF_model}{NMF object}

\item{cutoff}{Variant contribution cutoff to be used for database curreation}
}
\value{
Names list of gene and associated contribution
}
\description{
Function to calculate GSEA for each factor (this is running GO for some reason, where did I use this one at?)
}
\examples{
RNK_list <- factor_ranks(NMF_model = NMF_object, factor = 21)
}
