% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{calc_go_w}
\alias{calc_go_w}
\title{Gene Ontology Enrichment}
\usage{
calc_go_w(NMF_model, cutoff = 1e-04)
}
\arguments{
\item{NMF_model}{NMF_object}

\item{cutoff}{Contribution cutoff for variants to be used in GO enrichment}
}
\value{
Gene ontology data frame
}
\description{
Function to calc GO enrichment for each factor of the NMF model
}
\examples{
go_df <- calc_go_w(NMF_model = NMF_object)
}
