% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{cluster_dotplot}
\alias{cluster_dotplot}
\title{Hclust on the sNN object to create a dendrogram of sNN clusters.}
\usage{
cluster_dotplot(dimred_df, keyword, clusters, sNN, return_plotlist = F)
}
\arguments{
\item{dimred_df}{Dataframe of dimensional reduction of phenotype matrix}

\item{keyword}{Array of character strings to use for trait searching}

\item{clusters}{Cluster assignments for each trait (can use sNN clustering output)}

\item{sNN}{shared nearest neighbors object.}

\item{return_plotlist}{If false traits one combined plot, else if T individual plots are returned as a list.}
}
\value{
Updated dimred_df with new column to identify if traits realted to keyword.
}
\description{
Function to update dimred_df highlighting traits related to keyword()
}
\examples{
cluster_dotplot(dimred_df, keyword = c("BMI", "Type 2 diabetes", "diet"), clusters = clusters, sNN = sNN_object)
}
