% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{count_factors}
\alias{count_factors}
\title{Count Factors}
\usage{
count_factors(dimred_df, h_matrix)
}
\arguments{
\item{dimred_df}{Dataframe of dimensional reduction of phenotype matrix}

\item{h_matrix}{Phenotypic matrix (NMF_model$h)}
}
\value{
Dimred_df with new column of count values
}
\description{
Function to calulcate the number of factors with non zero weight for each phenotype
}
\examples{
count_factors(dimred_df, h_matrix = NMF_model$h)
}
