% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{create_ucsc_track}
\alias{create_ucsc_track}
\title{Create UCSC Tracks}
\usage{
create_ucsc_track(file, NMF_model, factor, name, description, window = 5000)
}
\arguments{
\item{file}{filename}

\item{NMF_model}{NMF_object}

\item{factor}{Specific factor name to format table for}

\item{name}{Track name}

\item{description}{Track description}

\item{window}{Base pair length of window size to create contribution skyline}
}
\value{
Wriets table to file
}
\description{
Function to format a table in UCSC compatible format
}
\examples{
create_ucsc_track(file = "Factor21.bedGraph", NMF_model = NMF_object, factor = 21, name = "Factor 21", description = "Contribution skyline for factor 21 in a 5000bp window.")
}
