% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{dimred_kw}
\alias{dimred_kw}
\title{Plot Trait Keyword}
\usage{
dimred_kw(dimred_df, keyword, clusters, h_matrix, exact = F)
}
\arguments{
\item{dimred_df}{Dataframe of dimensional reduction of phenotype matrix}

\item{keyword}{Array of character strings to use for trait searching}

\item{clusters}{Cluster assignments for each trait (can use sNN clustering output)}

\item{h_matrix}{Phenotypic matrix (NMF_model$h)}

\item{exact}{If false traits contianing keyword are selected, else if exact = T only traits that match keyword are selected.}
}
\value{
Updated dimred_df with new column to identify if traits realted to keyword.
}
\description{
Function to update dimred_df highlighting traits related to keyword()
}
\examples{
dimred_kw(dimred_df, keyword = c("BMI", "Type 2 diabetes", "diet"), h_matrix = NMF_model$h)
}
