% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{enrichment_calc}
\alias{enrichment_calc}
\title{Enrichment Calculation}
\usage{
enrichment_calc(peaks, database)
}
\arguments{
\item{peaks}{dataframe where one column is a list of peak names and the other column is a c("yes", "no") of if it a top peak}

\item{databse}{top X peak names of the tissue/cell type}
}
\value{
p values for enrichment
}
\description{
Function to calculate overepresentation enrichment
}
\examples{
r <- enrichment_calc(peaks, database)
}
