% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{factor_enrichments}
\alias{factor_enrichments}
\title{Factor Enrichments}
\usage{
factor_enrichments(
  NMF,
  filtered_closest,
  narrowpeaks,
  num_closest = 500,
  num_peaks = 1000,
  factors = NA
)
}
\arguments{
\item{NMF}{NMF object}

\item{filtered_closest}{dataframe of bedtools closest object}

\item{narrowpeaks}{dataframe of scATAC seq peak calling object}

\item{num_closest}{Number of closest peaks top use for enrichment calculation}

\item{num_peaks}{Number of top peaks to use for atac specificity and enrichment calc}

\item{factos}{Option to select only a specific set of NMF factors to calculate enrichment for}
}
\value{
Data frame of factors by cell type or tissue enrichment
}
\description{
Function to calculate overepresentation enrichment for each factor in NMF model
}
\examples{
enrichment_df <- factor_enrichments(NMF = NMF_object, filtered_closest, narrowpeaks)
}
