% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{factor_ranks}
\alias{factor_ranks}
\title{Factor Rank List}
\usage{
factor_ranks(NMF_model, factor)
}
\arguments{
\item{NMF_model}{NMF object}

\item{factor}{Name of factor that we want to create a ranked list for}
}
\value{
Names list of gene and associated contribution
}
\description{
Function to created a ranked list for gsea
}
\examples{
RNK_list <- factor_ranks(NMF_model = NMF_object, factor = 21)
}
