% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{factor_spider_plot}
\alias{factor_spider_plot}
\title{NMF Factor Udler Spider Plot}
\usage{
factor_spider_plot(factor, h_df, maxs, mins, custom_title = F)
}
\arguments{
\item{factor}{NMF factor}

\item{h_df}{Data.frame of the NMF h_matrix}

\item{maxs}{Max values for each of these traits}

\item{mins}{Min values for each of these traits}

\item{custom_title}{Option to set a specific title name for the plot}
}
\value{
spider plot for a specific nmf factor
}
\description{
Function to plot spider plots to compare with Udler bNMF paper.
}
\examples{
factor_spider_plot(factor = 21, h_df = NMF$h, maxs, mins)
}
