% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{fix_dups}
\alias{fix_dups}
\title{Fix Duplicates}
\usage{
fix_dups(go_df)
}
\arguments{
\item{go_df}{Gene ontology results dataframe}
}
\value{
Gene ontology data frame with duplicates removed
}
\description{
Function to fix the problem of merging creating replicates, need to find a proper way to merge that doesnt cause this
}
\examples{
go_df <- fix_dups(go_df)
}
