% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{format_factor_rank}
\alias{format_factor_rank}
\title{Format Factor Rank}
\usage{
format_factor_rank(factor_rank)
}
\arguments{
\item{factor_rank}{factor_rank}
}
\value{
Updated factor rank
}
\description{
Function to format factor rank
}
\examples{
factor_rank <- format_factor_rank(factor_rank)
}
