% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{format_go_df}
\alias{format_go_df}
\title{Format GO Dataframe}
\usage{
format_go_df(go_df, cutoff = 0.05)
}
\arguments{
\item{go_df}{Gene ontology dataframe, output of calc_go_w()}

\item{cutoff}{Significance cutoff for GO enrichments}
}
\value{
Gene ontology data frame with only significant enrichments
}
\description{
Function to format the output go_df
}
\examples{
go_df <- format_go_df(go_df)
}
