% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{get_nmf_genes}
\alias{get_nmf_genes}
\title{Get top Factor Genes}
\usage{
get_nmf_genes(NMF_model, factor, cutoff = 5e-05)
}
\arguments{
\item{NMF_model}{NMF object}

\item{factor}{Specific nmf factor}

\item{cutoff}{Contribution cutoff to find genes assocatiated to SNPs above this cutoff}
}
\value{
list of unique gene names
}
\description{
Function to find the genes associated with the top SNPs in a factor
}
\examples{
factor21_genes <- get_nmf_genes(NMF_model = NMF_object, factor = 21)
}
