% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{get_unique_GOterms}
\alias{get_unique_GOterms}
\title{Get Unique GO Enrichments}
\usage{
get_unique_GOterms(go_df, factor, pval_cutoff = 0.05)
}
\arguments{
\item{go_df}{Formatted GO enrichment dataaframe}

\item{factor}{Name of factor that we want specific enrichments for}

\item{pval_cutoff}{Significance cutoff for GO terms}
}
\value{
List of GO terms that are specifically significant in that factor only
}
\description{
Function to return a list of GO terms that are specifically significant in that factor only
}
\examples{
get_unique_GOterms(go_df, factor = 21)
}
