% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{highlight_clusters}
\alias{highlight_clusters}
\title{Highlight clusters.}
\usage{
highlight_clusters(dimred_df, keyword, h_matrix, clusters, exact = F)
}
\arguments{
\item{dimred_df}{Dataframe of dimensional reduction of phenotype matrix}

\item{keyword}{Array of character strings to use for trait searching}

\item{h_matrix}{Phenotypic matrix (NMF_model$h)}

\item{clusters}{Cluster assignments for each trait (can use sNN clustering output)}

\item{exact}{If false traits contianing keyword are selected, else if exact = T only traits that match keyword are selected.}
}
\value{
Updated dimred_df with new column to identify if traits realted to keyword.
}
\description{
Function to plot the clusters from the keywords to accompany the dotplot/hclut figure.
}
\examples{
highlight_clusters(dimred_df, keyword = c("BMI", "Type 2 diabetes", "diet"), h_matrix = NMF$h, clusters = clusters, sNN = sNN_object)
}
