% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{multiple_ridges}
\alias{multiple_ridges}
\title{Multiple Ridges}
\usage{
multiple_ridges(
  NMF_model,
  nmf_factors,
  window = 10000,
  chromosome,
  custom_range = NA
)
}
\arguments{
\item{NMF_model}{NMF object}

\item{nmf_factors}{Numerical positions or column names of individual factor}

\item{window}{Number of base pairs that the function will smooth the manhattan skyline over}

\item{chromosome}{Which chromosome to plot skyline over}

\item{custom_range}{Array of length 2 with start and stop positions along chromosome.}
}
\value{
A dataframe of window with corresponding max contribution in the window. To be used for plotting with plot_multiple_ridges()
}
\description{
Function to create skyline tracks of NMF factors
}
\examples{
# Calculating max contributions in factors 21, 54, and 108 at the NNAT locus (GRCh37) in 0.1 Mb windows
multiple_ridges_df <- multiple_ridges(NMF_model, nmf_factor = c(21, 54, 108), window = 100000, chromosome = 20, custome_range = c(36149617,36152092))
}
