% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{NMF_manhattan}
\alias{NMF_manhattan}
\title{NMF Manhattan}
\usage{
NMF_manhattan(NMF_model, nmf_factor, subset_w = NA)
}
\arguments{
\item{NMF_model}{NMF object}

\item{nmf_factor}{Numerical position or column name of individual factor}

\item{subset_w}{option to select a subset of the w matrix}
}
\value{
A ggplot formatted manhattan plot
}
\description{
Function to create NMF Factor manhattan plot
}
\examples{
plot <- NMF_manhattan(NMF_model, nmf_factor = 21)
}
