% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{only_sig}
\alias{only_sig}
\title{Plot Only Significant Enrichments}
\usage{
only_sig(go_df, factors)
}
\arguments{
\item{go_df}{Formatted GO enrichment dataaframe}

\item{factors}{Array of NMF factor integer names to plot}

\item{hclust_obj}{Option to provide an hclust object to cluster NMF factors}

\item{cutoff}{Variant contribution cutoff for GO enrichment calc}
}
\value{
Heatmap of significant GO enrichments across specified NMF factors
}
\description{
Function to Plot unique significant GO enrichments for specific NMF factors
}
\examples{
only_sig(go_df, factors = c(21, 87, 109))
}
