% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{pheno_hm}
\alias{pheno_hm}
\title{Phenotype Matrix Heatmap}
\usage{
pheno_hm(
  NMF,
  keyword,
  nmf_hc,
  out_dir,
  exact = F,
  custom_filename = F,
  column = "description"
)
}
\arguments{
\item{NMF}{NMF object}

\item{keyword}{Array of character strings to use for trait searching}

\item{nmf_hc}{Hclust object on the phenotypic matrix}

\item{out_dir}{Path to figure output directory}

\item{exact}{If false traits contianing keyword are selected, else if exact = T only traits that match keyword are selected.}

\item{custom_filename}{If false default name by keywords is used, if character string provided this will be used for file prefix}

\item{column}{Which column of the phenoitypic metadata to use for searching trait names}
}
\value{
Pheatmap of traits related to keyword columns ordered by hclust
}
\description{
Function to plot traits related to keyword on UMAP of the h matrix
}
\examples{
pheno_hm(NMF, keyword = c("BMI", "Type 2 diabetes", "diet"), nmf_hc, out_dir = "./path_to_output/")
}
