% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{plot_go_heatmap}
\alias{plot_go_heatmap}
\title{Plot GO Enrichment}
\usage{
plot_go_heatmap(NMF, filename, hclust_obj = F, cutoff = 1e-04)
}
\arguments{
\item{NMF}{NMF object}

\item{filename}{Name of file to save heatmap to}

\item{hclust_obj}{Option to provide an hclust object to cluster NMF factors}

\item{cutoff}{Variant contribution cutoff for GO enrichment calc}
}
\value{
Heatmap of GO enrichments across NMF factors
}
\description{
Function to plot a heatmap of GO enrichments
}
\examples{
plot_go_heatmap(NMF = NMF_object, filename = "/path_to_outoput/GO_heatmap.pdf")
}
