% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{plot_kw}
\alias{plot_kw}
\title{Plot Trait Keyword}
\usage{
plot_kw(
  dimred_df,
  keyword,
  h_matrix,
  dimred = "tSNE",
  size = "num_factors",
  clusters = F
)
}
\arguments{
\item{dimred_df}{Dataframe of dimensional reduction of phenotype matrix}

\item{keyword}{Array of character strings to use for trait searching}

\item{h_matrix}{Phenotypic matrix (NMF_model$h)}

\item{dimred}{Which type of dimensional reduction to plot}

\item{size}{Values to pass to ggplot to change size of poiints}

\item{clusters}{Cluster assignments for each trait (can use sNN clustering output)}
}
\value{
Ggplot of dimensional reduction highlighting the traits related to supplied keywords
}
\description{
Function to highlight traits related to an array of keywords on a dimensional reduction map.
}
\examples{
plot_kw(dimred_df, keyword = c("BMI", "Type 2 diabetes", "diet"), h_matrix = NMF_model$h)
}
