% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{plot_kw_umap}
\alias{plot_kw_umap}
\title{Plot keyword UMAP}
\usage{
plot_kw_umap(h_umap, keyword)
}
\arguments{
\item{h_umap}{UMAP output of h_matrix}

\item{keyword}{Array of character strings to use for trait searching}
}
\value{
Ggpoint plot of UMAP and highlighted traits
}
\description{
Function to plot traits related to keyword on UMAP of the h matrix
}
\examples{
plot_kw_umap(h_umap, keyword = c("BMI", "Type 2 diabetes", "diet"))
}
