% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{plot_multiple_ridges}
\alias{plot_multiple_ridges}
\title{Plot Multiple Ridges}
\usage{
plot_multiple_ridges(
  mult_ridge_df,
  squish = 10000,
  lwd = 1,
  alpha = 1,
  add_chr = F
)
}
\arguments{
\item{mult_ridge_df}{Dataframe of chromosome window with associated max factor contributions. Output of multiple_ridges()}

\item{squish}{Geom_ridgeline parameter to move stacked plots closer together by squishing the y axis}

\item{lwd}{Plot line width}

\item{alpha}{Transparency of plot color fill}

\item{add_chr}{Option to add ideogram at the bottom of the plot to show where this region is.}
}
\value{
A plot showing the skyline between factors across a specific region
}
\description{
Function to create skyline tracks of NMF factors
}
\examples{
# Calculating max contributions in factors 21, 54, and 108 at the NNAT locus (GRCh37) in 0.1 Mb windows
multiple_ridges_df <- multiple_ridges(NMF_model, nmf_factor = c(21, 54, 108), window = 100000, chromosome = 20, custome_range = c(36149617,36152092))
plot_mutliple_ridges(multiple_ridges_df)
}
