% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{sNN_clusters_plots}
\alias{sNN_clusters_plots}
\title{sNN Clusters Panel}
\usage{
sNN_clusters_plots(dimred_df, keyword, clusters, sNN, h_matrix)
}
\arguments{
\item{dimred_df}{Dataframe of dimensional reduction of phenotype matrix}

\item{keyword}{Array of character strings to use for trait searching}

\item{clusters}{Cluster assignments for each trait (can use sNN clustering output)}

\item{sNN}{shared nearest neighbors object.}

\item{h_matrix}{Phenotypic matrix (NMF_model$h)}
}
\value{
Updated dimred_df with new column to identify if traits realted to keyword.
}
\description{
Function to plot a panel of figures including highlighted traits, dendrogram of clusters, and dimensional reduction ggpoint of there.
}
\examples{
sNN_clusters_plots(dimred_df, keyword = c("BMI", "Type 2 diabetes", "diet"), h_matrix = NMF$h, clusters = clusters, sNN = sNN_object)
}
