% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{split_peak_name}
\alias{split_peak_name}
\title{Split peak name}
\usage{
split_peak_name(narrowpeaks, col_name = "name")
}
\arguments{
\item{narrowpeaks}{Narrowpeaks dataframe}

\item{col_name}{Name to pull from metadata}
}
\value{
DUpdated dataframe with formatted names
}
\description{
Function to clean up narrowpeak file names
}
\examples{
narrowpeaks <- split_peak_name(narrowpeaks)
}
