% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{test_perplexity}
\alias{test_perplexity}
\title{Test Perplexity}
\usage{
test_perplexity(NMF_model, perplexity, return_df = F)
}
\arguments{
\item{NMF_model}{NMF object}

\item{perplexity}{Perplexity value passed to Rtsne()}

\item{return_df}{Option to return plots or data frame of dimensional reduction}
}
\value{
A plot of dimensional reduction at the specified perplexity or a dataframe if return_df=T
}
\description{
Function to plot dimenstional reduction at different perplexities
}
\examples{
plot_p40 <- test_perplexity(NMF_model, perplexity = 40)
}
