% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{tissue_enrichment_df}
\alias{tissue_enrichment_df}
\title{Tissue Enrichment by GSEA}
\usage{
tissue_enrichment_df(NMF, closest, narrowpeaks, num_top_peaks = 100)
}
\arguments{
\item{closest}{path to bedtools closest output}

\item{narrowpeaks}{path to scATAC seqs narrowpeaks file}

\item{num_top_peaks}{Number of top peaks to use for tissue specificity from tissue peak calling}

\item{NMF_model}{NMF_object}
}
\value{
Dataframe of tissue enrichments for each factor
}
\description{
Function to calculate tissue enrichment by gsea
}
\examples{
calc_gsea_per_tissue(NMF = NMF_object, closes = "path_to_closest_file", narrowpeaks = "path_to_narrowpeaks_file")
}
