% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{trait_factor_weights}
\alias{trait_factor_weights}
\title{Plot Trait Weights}
\usage{
trait_factor_weights(NMF, trait)
}
\arguments{
\item{NMF}{NMF object}

\item{trait}{Specific name of a trait}
}
\value{
GGplot of a trait's weight across all factors
}
\description{
Function to plot a trait's weights across all factors
}
\examples{
trait_factor_weights(NMF = NMF_object, trait = "BMI")
}
