% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_utils.R
\name{write_tracks}
\alias{write_tracks}
\title{Write UCSC Tracks}
\usage{
write_tracks(out_dir, NMF_model, window = 5000)
}
\arguments{
\item{out_dir}{Path to output directory}

\item{NMF_model}{NMF_object}

\item{window}{Base pair length of window size to create contribution skyline}
}
\value{
Writes tracks for each of the NMF factors
}
\description{
Function to write UCSC tracks of NMF factors
}
\examples{
write_tracks(out_dir = "/path_to_output/", NMF_model = NMF_object)
}
